requirements.txt
absl-py==0.10.0
astunparse==1.6.3
backcall==0.2.0
bokeh==2.2.1
botocore==1.17.55
cachetools==4.1.1
cachey==0.2.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
contextvars==2.4
cycler==0.10.0
dask==2.20.0
decorator==4.4.2
distributed==2.25.0
docutils==0.15.2
Flask==1.1.2
fsspec==0.7.4
gast==0.3.3
gcsfs==0.6.2
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
HeapDict==1.0.1
hub==0.11.0.0
idna==2.10
immutables==0.14
importlib-metadata==1.7.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
littleutils==0.2.2
locket==0.2.0
lz4==3.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
msgpack==1.0.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
outdated==0.2.0
packaging==20.4
pandas==1.1.1
parso==0.7.1
partd==1.1.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
protobuf==3.13.0
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3fs==0.4.2
scipy==1.4.1
six==1.15.0
sortedcontainers==2.2.2
tblib==1.7.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.9.0
termcolor==1.1.0
toolz==0.10.0
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zict==2.0.0
zipp==3.1.0
